changed all fortran string length parameters to be size_t simplified fortran wrappers for windows
git-svn-id: svn://136.177.114.72/svn_GW/IPhreeqc/trunk@9015 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ IPhreeqc.cpp
@@ -874,7 +874,7 @@ void IPhreeqc::SetBasicCallback(double (*fcn)(double x1, double x2, const char *
 	this->PhreeqcPtr->register_basic_callback(fcn, cookie1);
 }
 
-void IPhreeqc::SetBasicFortranCallback(double (*fcn)(double *x1, double *x2, char *str, int l))
+void IPhreeqc::SetBasicFortranCallback(double (*fcn)(double *x1, double *x2, char *str, size_t l))
 {
 	this->PhreeqcPtr->register_fortran_basic_callback(fcn);
 }
